Syntactic Sugar & Style Guide Changes
Godot 4.5 added new syntactic sugar in the form of the abstract keyword. This makes overriding the function "execute" more explicit -- the compiler throws an error if it is not properly defined in an extended class. While not strictly necessary, it's good to follow the intended practice.
https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_basics.html#abstract-classes-and-methods
We also convert explicitly written NodePaths to exports, a change added in Godot 4.0, so the dependencies are less brittle (in case you want to move nodes around the SceneTree)
https://docs.godotengine.org/en/4.4/tutorials/scripting/gdscript/gdscript_exports.html#nodes

Some nitpicking changes to more closely follow the style guide:
1. Adding a space between comments and the hash (#)
https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_styleguide.html#comment-spacing
2. Explicitly writing ints as floats
https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_styleguide.html#numbers
3. Renaming randVec to _random_vector()
https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_styleguide.html#naming-conventions